pre-calibo-leave

+++ gitea-server/modified-entrypoint.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-if [ -f "/tmp/init-data-directory.tar.gz" ]; then
-    echo "Initial directory found. Moving it to data directory"
+if [ -f "/tmp/init-data-directory.tar.gz" ] && [ ! -f /data/movedflag ]; then
+    echo "Initial directory and flag found. Moving it to data directory"
     cd /tmp
     tar -xzvf init-data-directory.tar.gz
     mv /tmp/gitea/* /data
